fix build errors and lint
